pyproject.toml
[project]
name = "employeetask"
version = "0.1.0"
description = ""
authors = [
    {name = "Dima",email = "gruzdev-dmitriy34@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "ipython (>=9.6.0,<10.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "coverage (>=7.10.7,<8.0.0)",
    "stripe (>=13.0.1,<14.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.9.0"
isort = "^7.0.0"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "config/asgi.py",
    "config/wsgi.py",
    ]
